Functional 16bit text mode

+++ src/videomodes.h
@@ -1,2 +1,6 @@
+#include <stdint.h>
+#include "memorymap.h"
+#include "globals.h"
+
 void videoSetSlideMode(void);
-void videoSetActiveSlideBuffer(unsigned bufferId);
+void videoSetActiveSlideBuffer(uint8_t bufferId);
